updated the following with this function:
(a) quantified the density of air given temp and pressure

(b) acquired the product of ppmw and the density of air

(c) the concentrations are quantified as mg / m-3 for both gases

data has been exported to an  file

this commit satisfies four of the five tasks highlighted in issue #5. however, since the first task, converting temps and pressures is completed near the top of each `.m` script, this commit closes #5.