major updates that resolved issue #7
edits to the  file found that the '-2' file in the 11Jul2017 under 'VCAC' was the troublesome file. after deleting the last row of the file, the first task
and second tasks of #7 appear to be resolved

the third task of the issue #7, also appears to be resolved after executing the entire `.m` scripts for the valles sites visited in 2017 successfully.

finally, the fourth task of #7 was completed successfully after manually copying the edited `dat` files to the OneDrive directory

therefore, issue #7 can be considered closed (close #7)