MATLAB-Valles (#8)
**Overall**, I completed the work necessary to update the flux calculations with the new method (i.e., `gas_unit_prep()`). 

* major updates that resolved issue #7

edits to the  file found that the '-2' file in the 11Jul2017 under 'VCAC' was the troublesome file. after deleting the last row of the file, the first task
and second tasks of #7 appear to be resolved

the third task of the issue #7, also appears to be resolved after executing the entire `.m` scripts for the valles sites visited in 2017 successfully.

finally, the fourth task of #7 was completed successfully after manually copying the edited `dat` files to the OneDrive directory

therefore, issue #7 can be considered closed (close #7)

* the 10Jul2017  script runs successfully after organizing the data in the  dir

* successfully updated and ran the valles sites except for the 25may2019 because that one is based off of h5 files

* successfully used the original template for the 25May2019 (VCSS) site

however, I had to update the accompanying function  to get the script to work. it is unlikely that this one will be updated with the rest because it is unique since it uses  files instead of  files. thus, an addition to the  should be considered

* completed some more moderate updates to ensure that the 25May2019 (VCSS) was using the updated method to calculate functions

also, committing the 08Jul2017 file now that all of the valles data has been recalculated

this commit should end the branch 'matlab-valles' and merge back into 'matlab'

---------

Co-authored-by: 907Resident <907Resident@users.noreply.github.com>