Matlab (#10)
**Overall** notes: see this [comment](https://github.com/907Resident/caldera-gases/pull/10#issue-1641064217) for summary of work completed in this PR

* initial commit for the data viewer web app

more files will be added but this new branch will be used to develop the initial web app product

* updated the ignore file to omit __pycache__ files

* added a rough draft of the opening page for the web app

this still needs work. I have only resolved two of the four checkmarks for issue #3. the resolved ones are: (1) a draft for the title of app and (2) a description of the project and how the app is related to the project as a whole

* added a rough draft of the opening page for the web app

the second map, choropleth, is preferred. however some work on the hover labels is required

further thought on the rest of the maps/charts in this section is required as well

* initial commit for calderas package

this package still needs more modules but will be useful when attempting to modularize the code in this project

* adding a .keep file to maintain structure

* moved the site-differentation notebook on this branch. should be care with that on the main

* moved the matlab-functions dir under a new dir, 'matlab-functions'

* initial commit of .m files

I added many of the .m files that were originally used to calculate fluxes

* drafted changes to the opening section where path variables are set so that the paths are relative instead of absolute

still testing this to ensure it works properly

* updated the opening section and added a 'site_metadata'  file to assist with running the entire  script

* initial commit

added all of the MATLAB functions needed to execute flux calculation. needs to be verified

* updated the  function

this update was built to satisfy issue #6. with this commit the first draft to fix the file paths has been completed. a couple more checks are required to complete this issue

* updated the ignore file to ignore the data in the non-pertinent-data directory

* updated these files to allow for the computation (last10) and plotting (isocomp) of the isotopic data after the update for the 'new' flux calculation

* updated the following with this function:

(a) quantified the density of air given temp and pressure

(b) acquired the product of ppmw and the density of air

(c) the concentrations are quantified as mg / m-3 for both gases

data has been exported to an  file

this commit satisfies four of the five tasks highlighted in issue #5. however, since the first task, converting temps and pressures is completed near the top of each `.m` script, this commit closes #5.

* made major updates to the PRCN file and using it as a template to successfully update the flux calcualtions as described in #5.

the remaining files will use this scheme unless determined otherwise

* moving files around

* attempting to right a bad deletion and ocmmit

* ensuring that the custom functions are added to the working tree

* also ensuring that the functions are added to the working tree

* restored a previously deleted  file from Google Drive trash

* adding back some more files that were deleted

more to come

* adding back more deleted files to working tree

* another restored file added back

* adding a .pdf that explains the _new_ method for converting the obsverations from ppm to mass per area per hour

* collecting changes before making a valles specific branch

08Jul2017 ran after some work

* MATLAB-Valles (#8)

**Overall**, I completed the work necessary to update the flux calculations with the new method (i.e., `gas_unit_prep()`). 

* major updates that resolved issue #7

edits to the  file found that the '-2' file in the 11Jul2017 under 'VCAC' was the troublesome file. after deleting the last row of the file, the first task
and second tasks of #7 appear to be resolved

the third task of the issue #7, also appears to be resolved after executing the entire `.m` scripts for the valles sites visited in 2017 successfully.

finally, the fourth task of #7 was completed successfully after manually copying the edited `dat` files to the OneDrive directory

therefore, issue #7 can be considered closed (close #7)

* the 10Jul2017  script runs successfully after organizing the data in the  dir

* successfully updated and ran the valles sites except for the 25may2019 because that one is based off of h5 files

* successfully used the original template for the 25May2019 (VCSS) site

however, I had to update the accompanying function  to get the script to work. it is unlikely that this one will be updated with the rest because it is unique since it uses  files instead of  files. thus, an addition to the  should be considered

* completed some more moderate updates to ensure that the 25May2019 (VCSS) was using the updated method to calculate functions

also, committing the 08Jul2017 file now that all of the valles data has been recalculated

this commit should end the branch 'matlab-valles' and merge back into 'matlab'

---------

Co-authored-by: 907Resident <907Resident@users.noreply.github.com>

* completed the recalculations for the yellowstone fluxes (#9)

**Overall, I have finished off the changes required to recalculate the Yellowstone fluxes using the new method present in the code for `gas_unit_prep`. See commit messages below for further details.

edited the 06Jun2018, 14Jun2019, and 12Jun2019 files to finish off the edits required to ensure the data could be recalculated

much of the work that was required to complete the yellowstone recalculations was originally completed on the 'matlab' branch. thus, a reader of this commit should look at 372ced1 as a key example of what needed to be done to begin the changes

Co-authored-by: 907Resident <907Resident@users.noreply.github.com>

* committing aux files before completing PR #10

---------

Co-authored-by: 907Resident <907Resident@users.noreply.github.com>